+++ darglint/driver.py
@@ -17,12 +17,18 @@ from .config import (
 
 parser = argparse.ArgumentParser(description='Check docstring validity.')
 parser.add_argument(
-    '--verbosity',
-    '-v',
-    default=1,
-    type=int,
-    choices=[1, 2],
-    help='The level of verbosity.',
+    '--message-template',
+    '-m',
+    type=str,
+    help=(
+        'Specify a message template.  This is a Python format string '
+        'describing errors, which can access the following attributes:\n'
+        '    line: The line number,\n'
+        '    msg: The error message,\n'
+        '    msg_id: The error code,\n'
+        '    obj: The function/method name,\n'
+        '    path: The relative file path.\n'
+    ),
 )
 parser.add_argument(
     '--raise-syntax',
@@ -31,6 +37,14 @@ parser.add_argument(
         'When a docstring is incorrectly formatted, raise an exception '
         'rather than storing the error.'
     ),
+)
+parser.add_argument(
+    '--verbosity',
+    '-v',
+    default=1,
+    type=int,
+    choices=[1, 2],
+    help='The level of verbosity.',
 )
 parser.add_argument(
     'files',
@@ -44,8 +58,13 @@ parser.add_argument(
 # ---------------------- MAIN SCRIPT ---------------------------------
 
 
-def get_error_report(filename, verbosity, config, raise_errors_for_syntax):
-    # type: (str, int, Configuration, bool) -> str
+def get_error_report(filename,
+                     verbosity,
+                     config,
+                     raise_errors_for_syntax,
+                     message_template=None,
+    ):
+    # type: (str, int, Configuration, bool, str) -> str
     """Get the error report for the given file.
 
     Args:
@@ -56,6 +75,8 @@ def get_error_report(filename, verbosity, config, raise_errors_for_syntax):
             to propagate up (crashing darglint.)  This is useful
             if we are developing on darglint -- we can get the stack
             trace and know exactly where darglint failed.
+        message_template: A python format string for specifying
+            how the message should appear to the user.
 
     Returns:
         An error report for the file.
@@ -70,7 +91,11 @@ def get_error_report(filename, verbosity, config, raise_errors_for_syntax):
     )
     for function in functions:
         checker.run_checks(function)
-    return checker.get_error_report(verbosity, filename)
+    return checker.get_error_report(
+        verbosity,
+        filename,
+        message_template=message_template,
+    )
 
 
 def main():
@@ -84,13 +109,13 @@ def main():
     args = parser.parse_args()
     files = [x for x in args.files if x.endswith('.py')]
     raise_errors_for_syntax = args.raise_syntax or False
-    verbosity = args.verbosity
     for filename in files:
         error_report = get_error_report(
             filename,
-            verbosity,
+            args.verbosity,
             config,
-            raise_errors_for_syntax
+            raise_errors_for_syntax,
+            message_template=args.message_template,
         )
         if error_report:
             print(error_report + '\n')