+++ darglint/integrity_checker.py
@@ -272,14 +272,16 @@ class IntegrityChecker(object):
             self.errors.sort(key=lambda x: x.function.lineno)
             self._sorted = True
 
-    def get_error_report(self, verbosity, filename):
-        # type: (int, str) -> str
+    def get_error_report(self, verbosity, filename, message_template=None):
+        # type: (int, str, str) -> str
         """Return a string representation of the errors.
 
         Args:
             verbosity: The level of verbosity.  Should be an integer
                 in the range [1,3].
             filename: The filename of where the error occurred.
+            message_template: A python format string for describing
+                how the error reports should look to the user.
 
         Returns:
             A string representation of the errors.
@@ -289,4 +291,5 @@ class IntegrityChecker(object):
             errors=self.errors,
             filename=filename,
             verbosity=verbosity,
+            message_template=message_template,
         ))